Task tree path names now available.
This required the introduction of a slightly special case around the
root node again... for name handling, of all things... tsk.
This seems fairly acceptable; we will be needing it for panic recovery
to be consistent as well (not... that I expect this to be a huge fuss,
since if the first think you're doing is launching *more supervisors*
(and it almost certainly is), and the supervisors are pretty well known
not to panic unless something is supercritically wrong and you should
probably just halt the program... BUT STILL).

Signed-off-by: Eric Myhre <hash@exultant.us>