Fix for and test example for cancellation path.
Writing a test for concurrent behavior which doesn't add so many
memory fences that it removes the load from the system under test
is extremely hard.  It's not that we can't put more mutexes and
channels and so forth in the system to force its behavior; it's
more that we can't do it without forcing more synchronization than is
*normal* for the supervisor logic, and if it overlays any of the
synchronization we're trying to *test*, then it actually makes the
test useless and trivial.  So, while I'm not satisfied by the use of
sleep calls here... improvements without trivializing the test are
hard to see.

Signed-off-by: Eric Myhre <hash@exultant.us>