supervisorStream implementation; generalizations.
supervisorStream supports the use of TaskGen channels to produce an infinite
set of tasks.  If the channel is closed, the supervisor will continue to wait
for all currently submitted and in-progress tasks to finish; if the supervisor
is cancelled, it will behave much the same as supervisorForkJoin when cancelled
(fanning out cancel to child tasks, but again still awaiting them), and the
TaskGen channel will no longer be serviced.

The supervisorStream implementation is quite a bit nicer than the supervisorFJ
implementation.  Each phase of execution gets its own function, which means
if we get a process-halting panic (or other all-stacks-logging event), we'll
easily be able to see what phase each supervisor was in at the time.

And now that we have two full implementations, we can reasonably see where
some things can now be called common shared beheaviors...

Turns out the "launchChild" functions didn't really need any reference to the
supervision manager objects at all, as long as it's the case that the report
channel pattern really is in common.  And the report channel pattern really is
in common between our two implementations now.  So the launchChild function,
the report channel pattern, and its message type are now moved to a 'shared'
file.

All the phase markers also moved to the shared file.  These phase markers serve
very little practical purpose at present, it should be admitted; really they're
there for blocking multiple calls to `Run`, and the rest... my aim is to make
it available for monitoring and logging, but that's about the most you can ever
expect from the phase info.  (Trying to make any programmatic logic from the
happens-after view of the phase info would only lead you odd, dark places).

A TaskGenFromTasks factory method has appeared.  It's a shape you probably
won't often need to use (pretending a finite set of already-known tasks is
actually a TaskGen channel), but it's there.

The public SuperviseForkJoin method now quietly uses the new supervisorStream
implementation Because It Can.  This will probably switch back; for now, it's
a quick cheat to make our existing tests prove it works (but of course really
we should be making more tests rather than pure examples ASAP).

Signed-off-by: Eric Myhre <hash@exultant.us>