Update supervisorFJ to atomics and phase funcs.
Behaviorally, this is identical; but now we're getting the same benefits
(monitorability in stack dumps, etc) as supervisorStream has.

Interesting, the collect and halting phases are... identical.  They could be
DRY'd entirely.  It's just init, init state, and the running phase that
are disinct now.  I'm not sure I can come up with any meaningful reason to go
on a refactor craze, though -- not until the code in question either gets more
complex, or there's a third instance brewing.  Just imagine a macro there.

Signed-off-by: Eric Myhre <hash@exultant.us>