Forkjoin supervisors are immediately "collecting".
The docs for phase_running say that new tasks can be submitted; that's never
true for a supervisorFJ.  It's also true that the phase info is explicitly
purely advisory (e.g. you'd be a maniac to look at the phase, see that it's
momentarily "running" and then assume you can actually submit a task more than
zero quantums of time later), but nonetheless.

Signed-off-by: Eric Myhre <hash@exultant.us>