Support panic recovery; and ErrChild type.
ErrChild will likely continue to grow: at least reporting the CtxTaskPath seems
extremely useful.

There's always the argument to be had about concatenative context and what
order to do it in: should we report the whole CtxTaskPath immediately?  Or
build it incrementaly, by prepending CtxTaskName each time we rise a tree
level?  Either choice is irritating in some user story.

Added a test that exercises this, and the sheer fact it doesn't die in panic is
proof of something, at least.  We need an assertion library in here before I'm
going to make anything more fancy, though.

Signed-off-by: Eric Myhre <hash@exultant.us>