Promises.
I want to judicious about what goes into go-sup; there are many ways
to structure concurrency, and I want to provide composable primitives
rather than opinionated abstractions.

This is often further aided by the fact that Golang channels are pretty
awesome in the first place: most task submission and queuing signalling
concepts are well-presented by either channels alone or standard
library features like sync.WaitGroup and sync.Mutex.

Similarly, cancellation is already well-described by the standard
library context.Context system.

Promises still stand out, however.  Specifically, in the scenario where
it's desirable to wait on task results individually, the tasks are
unique in some way which makes bulk selection on results need further
associations, and using a closure to capture those associations is
unergonomic.  It's hard to describe rules of thumb for concurrent
design, but these situations do seem to come up regularly.

This is a somewhat different take on promises than the usual.  Notice
that the design is such that you *create* a promise, and *pass it down*
to the code that's then doing work.  This is in contrast with a more
historically stereotypical design of promises as a value you *return*.

This design is similar to the reasoning behind designing Go APIs to
accept channels are arguments rather than returning them: it allows the
*caller* to chose the appropriate level of reporting clearly, and wire
all handling before starting the task.

The Promise system implemented here is entirely optional to use.
No part of the Supervisor system will use or expose Promise types.
However, if you want to use Promise to build your own tasks or in
closures that become tasks... here's your tool.

Signed-off-by: Eric Myhre <hash@exultant.us>