api v4 drafts.
This repo as a whole is roughly v3 -- the v1 and v2 happened in a
separate root of history.

v4 is actually coming halfway back to v2, because the learn'in there
about the "go relevant.line.numbers.matter()" is still valid; but we're
also still going to keep the relationship with Context developed in v3,
which hopefully will result in the best of all worlds.

Some of these docs have been lurking/brewing for... a few weeks now,
and so at this point I'm just going to commit them as a checkpoint
before any more time slips away.  Implementation work is pending.
I'm a bit dissatisfied with the naming comments at the end still, and
that might need some more iteration.

Signed-off-by: Eric Myhre <hash@exultant.us>