Revisiting the supervision problem.
It's clear and studied and well-known that CSP can "do anything",
but when working with large amounts of golang code,
it becomes equally clear that _having some conventions helps_.
Legibility, reviewability, debuggability, and the sheer odds of
writing correct code in the first place -- all can drop rapidly
when writing a nontrivial application requires large amounts of
boilerplate, and requires great diligence in never forgetting
to e.g. check for cancellation.

And writing a nontrivial concurrent application in golang **does**
require large amounts of boilerplate, and large amount of that
dangged rare and difficult "diligence" thing.

So: maybe let's have a supervision library, which standardizes
some extremely common patterns wherever the type system allows us
to do so; document some consistent recommendations for whatever
the type system _can't_ help us enforce; and offer helper features
as library functions for whatever's in between and for anything
that can reduce sheer boilerplate.  Along the way, let's also
try to emit warnings for any situations that seem bad, and provide
as much support for debuggability and introspections as we can.

And this is my attempt.

(I've made a few shots at this before, and might combine all
their history by the time I publish this, but this is a new
root of git history for now, and a generally fresh take on it.
Compared to my own previous attempts, this one is focusing a
good bit more on simplifying what's made visible to the user;
it's full of "one-stop-shop" interfaces that are meant to feel
like an expert system you can just have a conversation with (rather
than breaking things down and exposing the entire problem domain
to the caller, which is how some of my previous shots did work!);
and it's also being written after golang has grown to have both
the Context type in the standard library, as well as the
type parameters (aka "generics") feature shipped in the compiler,
both of which shift the possibility space considerably.)

This commit is all interfaces (and all subject to change!), but
I think should get pretty close to the mark.