Believe this is a Promise implementation.
I've taken a pretty simple approach here.  It controls one value;
and it just offers a LOT of different ways to wait.

I've seen many other promise systems proposed since type parameters
(aka generics) came out in golang, and many of them give innate
blessing to a dichotomy of "resolve" (with a value) or
to "reject" (with an error value), and...
for better or for worse, I've decided to simply completely ignore that.
Values are values, and errors are values too.
The distinction does not seem useful.
When using this Promise type, one can define whatever values or errors
one wishes, including doing both together inside some other structure,
according to whatever distinctions are meaningful to you.
(I'm not really anticipating using this Promise type for making
complex chains of '.AndThen' callbacks, because we tend not to need
that in golang, so having "the railroad" of control flow split if
an error is seen has even less value than it otherwise would.)