various.
Prancing around implementing supervisor guts.  A few more features have
emerged along the way.

The main thing I'm concerned about in the current approach is the
double appearance of Context when creating and then when launching
a Supervisor.  If I could find a way out of that, I'd be happy.
However, as discussed in the new content amended to the README,
it really does seem unavoidable without compromising usability
in other much more critical ways.