First working pass of entire supervisor.Run implementation, including halting.  tasktree demo now runs entirely!
I'm definitely not confident that some of the internal code is the
simplest that it could be.  However, A) it works, and B) I actually
am pretty happy hewing to designs that emphasize stack trace legibility
even at the cost of some (near-)duplicated code.  When debugging,
or looking at all the stacks when the process dies to SIGQUIT,
the stack traces indicating which phase various systems are in really
is turning out to be consistently very pleasing.