fix task name tree growth.
Also fix typo in the tasktree example that made it more boring
than it was supposed to be... and slow it down to be more visible.

The placeholder names for there's-no-supervision-yet end up in the name
of the root supervisor right now.  That's probably silly.
I guess we ought to have a different constructor for a root supervisor,
so that you can give it a clean name.

Normally I'm opposed to package-scope variables, but maybe in this
case we should actually make a map for root supervisor names just so we
can enforce noncollision there.